# Implement client-side data persistence for instant app reopening