--- a/apps/mobile/package.json
+++ b/apps/mobile/package.json
@@ -18,7 +18,10 @@
   "dependencies": {
     "@clerk/clerk-expo": "^2.14.30",
     "@expo/vector-icons": "^15.0.2",
+    "@react-native-async-storage/async-storage": "^2.2.0",
+    "@tanstack/query-async-storage-persister": "^5.90.2",
     "@tanstack/react-query": "^5.87.4",
+    "@tanstack/react-query-persist-client": "^5.90.2",
     "@zine/design-system": "workspace:*",
     "@zine/shared": "workspace:*",
     "expo": "~54.0.10",